[#42] Automtic PID file path definition
In the case the user does not set the `pidfile` configuration
parameter, the software should define one path itself.
In particular the change makes the `pidfile` equal to the concatenation
of the `unix_socket_dir` (mandatory) and the string `pgmoneta.<host>.pid`
where `<host>` is the value of the `host` configuration parameter
or `all` if `host = *`.

Changed the `pidfile` variable size.

Updated the documentation and the AUTHORS file.

Close #42.